FEATURE: Moving gym import to __init__ for a shorter import statement. Pinning back the gym version.
