+++ docs/source/tutorials.rst
@@ -25,7 +25,7 @@ Example on how to use an environment e.g. send random movements
 
 .. code-block:: python
 
-    from myosuite.utils.import_utils import gym
+    from myosuite.utils import gym
     env = gym.make('myoElbowPose1D6MRandom-v0')
     env.reset()
     for _ in range(1000):
@@ -42,7 +42,7 @@ Example on how to generate and visualize a movement e.g. index flexion, and visu
 
 .. code-block:: python
 
-    from myosuite.utils.import_utils import gym
+    from myosuite.utils import gym
     env = gym.make('myoHandPoseRandom-v0')
     env.reset()
     for _ in range(1000):
@@ -58,7 +58,7 @@ Example on using a policy e.g. elbow flexion, and change non-stationaries
 
 .. code-block:: python
 
-    from myosuite.utils.import_utils import gym
+    from myosuite.utils import gym
     policy = "iterations/best_policy.pickle"
 
     import pickle
@@ -80,7 +80,7 @@ This example shows how to add fatigue to a model. It tests random actions on a m
 
 .. code-block:: python
 
-    from myosuite.utils.import_utils import gym
+    from myosuite.utils import gym
     env = gym.make('myoElbowPose1D6MRandom-v0')
     env.reset()
     for _ in range(1000):
@@ -104,7 +104,7 @@ This example shows how to add sarcopenia or muscle weakness to a model. It tests
 
 .. code-block:: python
 
-    from myosuite.utils.import_utils import gym
+    from myosuite.utils import gym
     env = gym.make('myoElbowPose1D6MRandom-v0')
     env.reset()
     for _ in range(1000):
@@ -129,7 +129,7 @@ This example shows how load a model with physical tendon transfer.
 
 .. code-block:: python
 
-    from myosuite.utils.import_utils import gym
+    from myosuite.utils import gym
     env = gym.make('myoHandKeyTurnFixed-v0')
     env.reset()
     for _ in range(1000):
@@ -165,7 +165,7 @@ If you want to load and execute the pre-trained DEP-RL baseline. Make sure that 
 
 .. code-block:: python
 
-    from myosuite.utils.import_utils import gym
+    from myosuite.utils import gym
     import deprl
 
     # we can pass arguments to the environments here
