+++ myosuite/envs/env_variants.py
@@ -5,7 +5,7 @@ Source  :: https://github.com/vikashplus/robohive
 License :: Under Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 ================================================= """
 
-from myosuite.utils.import_utils import gym; register=gym.register
+from myosuite.utils import gym; register=gym.register
 import collections
 from copy import deepcopy
 from flatten_dict import flatten, unflatten
