+++ myosuite/envs/myo/myobase/obj_hold_v0.py
@@ -5,7 +5,7 @@ Authors  :: Vikash Kumar (vikashplus@gmail.com), Vittorio Caggiano (caggiano@gma
 
 import collections
 import numpy as np
-from myosuite.utils.import_utils import gym
+from myosuite.utils import gym
 
 from myosuite.envs.myo.base_v0 import BaseV0
