+++ myosuite/envs/myo/myobase/reorient_sar_v0.py
@@ -5,7 +5,7 @@ Authors  :: Cameron Berg (cameronberg@meta.com), Vikash Kumar (vikashplus@gmail.
 
 import collections
 import numpy as np
-from myosuite.utils.import_utils import gym
+from myosuite.utils import gym
 
 from myosuite.envs.myo.base_v0 import BaseV0
 from myosuite.utils.quat_math import euler2quat, mulQuat, negQuat, mat2quat
