+++ myosuite/envs/myo/myochallenge/__init__.py
@@ -1,4 +1,4 @@
-from myosuite.utils.import_utils import gym; register=gym.register
+from myosuite.utils import gym; register=gym.register
 import os
 curr_dir = os.path.dirname(os.path.abspath(__file__))
 import numpy as np
