+++ myosuite/envs/myo/myochallenge/baoding_v1.py
@@ -5,7 +5,7 @@ Authors  :: Vikash Kumar (vikashplus@gmail.com), Vittorio Caggiano (caggiano@gma
 
 import collections
 import enum
-from myosuite.utils.import_utils import gym
+from myosuite.utils import gym
 import numpy as np
 
 from myosuite.envs.myo.base_v0 import BaseV0
