+++ myosuite/envs/myo/myodm/__init__.py
@@ -1,4 +1,4 @@
-from myosuite.utils.import_utils import gym; register=gym.register
+from myosuite.utils import gym; register=gym.register
 
 import collections
 import os
