+++ myosuite/envs/myo/myodm/myodm_v0.py
@@ -5,7 +5,7 @@ Source  :: https://github.com/MyoHub/myosuite
 License :: Under Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 ================================================= """
 
-from myosuite.utils.import_utils import gym
+from myosuite.utils import gym
 from myosuite.envs import env_base
 from myosuite.logger.reference_motion import ReferenceMotion
 from myosuite.utils.quat_math import quat2euler, euler2quat, quatDiff2Vel, mat2quat
