+++ myosuite/utils/import_utils.py
@@ -5,16 +5,6 @@ from os.path import expanduser
 import git
 
 
-# Utility to import gym/gymnasium
-def import_gym():
-    if importlib.util.find_spec("gymnasium"):
-        import gymnasium as gg
-    elif importlib.util.find_spec("gym"):
-        import gym as gg
-    return gg
-gym = import_gym()
-
-
 def mujoco_py_isavailable():
     help = """
         Options:
@@ -32,7 +22,6 @@ def mujoco_isavailable():
         (1) install robohive with encoders (pip install robohive['mujoco'])
         (2) follow setup instructions here: https://github.com/deepmind/mujoco
         (3) install mujoco via pip (pip install mujoco)
-
     """
     if importlib.util.find_spec("mujoco") is None:
         raise ModuleNotFoundError(help)
